Task: Added permissions column to user table for admin privilages, added logic to check for admin permission before showing Admin Menu and preventing user from Admin functionality if they are not admin.

+++ app/Http/Controllers/AdminController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 class AdminController extends Controller
 {
     /**
@@ -20,19 +21,36 @@ class AdminController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
+    public function index(){
+        //Check if the current user is admin
+        $currentUser = Auth::user();
+        if($currentUser->permission!=1){ // It is not admin and redirect to home
+            return redirect()->route('home')->withErrors(["Sorry, you're not the Administrator"]);
+        }
+
         $users = User::all();
         return view('admin')->with('users', $users);
     }
 
     public function deleteUser($userId){
+        //Check if the current user is admin
+        $currentUser = Auth::user();
+        if($currentUser->permission!=1){ // It is not admin and redirect to home
+            return redirect()->route('home')->withErrors(["Sorry, you're not the Administrator"]);
+        }
+
         $user = User::find($userId);
         $user->delete();
         return redirect()->route('admin')->with('message', 'User has been deleted!');
     }
 
     public function lockUser($userId){
+        //Check if the current user is admin
+        $currentUser = Auth::user();
+        if($currentUser->permission!=1){ // It is not admin and redirect to home
+            return redirect()->route('home')->withErrors(["Sorry, you're not the Administrator"]);
+        }
+
         $user = User::find($userId);
         $user->locked = 1;
         $user->save();
@@ -40,9 +58,16 @@ class AdminController extends Controller
     }
 
     public function unlockUser($userId){
+        //Check if the current user is admin
+        $currentUser = Auth::user();
+        if($currentUser->permission!=1){ // It is not admin and redirect to home
+            return redirect()->route('home')->withErrors(["Sorry, you're not the Administrator"]);
+        }
+
         $user = User::find($userId);
         $user->locked = 0;
         $user->save();
         return redirect()->route('admin')->with('message', 'User has been unlocked!');
     }
+
 }
